move CommitDiff types into fiber-types
